requirements.txt
pandas
requests
google-cloud-bigquery
google-cloud-storage